concurrency + bug fix

+++ proj-src/server/models/workouts.js
@@ -10,55 +10,50 @@ const { ObjectId, connect } = require('./mongo');
  * @property {number} privacy - 0 = public, 1 = private
  */
 
-/**
- * 
- * @typedef {Object} HasId
- * @property {string} _id - The ID of the document.
- */
-
 const COLLECTION_NAME = 'workouts';
 async function getCollection() {
     const db = await connect();
     return db.collection(COLLECTION_NAME);
 }
 
-/**
- * @returns {Promise<Workout[]>}
- */
 async function getAll() {
     const col = await getCollection();
     return col.find({}).toArray();
 }
 
-/**
- * @param {string} id - Index id.
- */
 async function get(id) {
+    if(!ObjectId.isValid(id)) return "Invalid ID";
+
     const col = await getCollection();
-    return col.findOne({ _id: new ObjectId(id)});
+    const workout = await col.findOne({ _id: new ObjectId(id) });
+
+    if(!workout) return "Workout not found";
+
+    return workout
 }
 
 async function getByName(name) {
     const col = await getCollection();
-    return col.findOne({ name: name });
+    return await col.findOne({ name: name });
 }
 
 async function add(workout) {
     const col = await getCollection();
-    return col.insertOne(workout);
+    return await col.insertOne(workout);
 }
 
 async function update(id, workout) {
-    const col = await getCollection();
-    // Remove the id
+    if(!ObjectId.isValid(id)) return null;
 
+    const col = await getCollection();
     delete workout._id;
     return col.updateOne({ _id: new ObjectId(id) }, { $set: workout });
 }
 
 async function remove(id) {
+    if(!ObjectId.isValid(id)) return null;
+
     const col = await getCollection();
-    console.log(id)
     return col.deleteOne({ _id: new ObjectId(id) });
 }
